sys: Enable -Wall, -Wextra, -Werror and fix all warnings on all Platforms (#483)
* sys: Make ImHex compile with -Wall -Wextra -Werror

* sys: Fixed various build errors on Linux

* sys: Explicitly ignore return value of `system` function

* sys: More fixes for the warnings GitHub Actions enables somehow

* sys: More fixes

* sys: Remove -Werror again to see all GitHub Actions warnings

* sys: Hopefully fixed all remaining warnings

* sys: Added back -Werror

* git: Change windows icon in GitHub Actions

+++ lib/libimhex/source/pattern_language/preprocessor.cpp
@@ -215,7 +215,7 @@ namespace hex::pl {
             });
 
             for (const auto &[define, value, defineLine] : sortedDefines) {
-                i32 index = 0;
+                size_t index = 0;
                 while ((index = output.find(define, index)) != std::string::npos) {
                     output.replace(index, define.length(), value);
                     index += value.length();
